solved ticket https://issues.jboss.org/browse/RAILO-197

+++ railo-java/railo-core/src/railo/commons/lang/KeyGenerator.java
@@ -8,7 +8,7 @@ public class KeyGenerator {
 		// create a crossfoot of the string and change result in constealltion of the position
 		long sum=0;
 		for(int i=value.length()-1;i>=0;i--){
-			sum+=((long)value.charAt(i))*((i%3+1)/2f);
+			sum+=(value.charAt(i))*((i%3+1)/2f);
 		}
 		return Md5.getDigestAsString(value)+":"+sum;
 	}
